refactor: move sub components into single files
